requirements.txt
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
xgboost==2.1.3
betfairlightweight==2.20.1
requests==2.32.3
aiohttp==3.10.10
understat==0.1.14
matplotlib==3.9.2
seaborn==0.13.2
scipy==1.13.1
pyarrow==17.0.0
python-dotenv==1.0.1
joblib==1.4.2
tqdm==4.67.1
